[Decomp][World] Cohesion split: extract session management to WorldSessionMgr.cpp
